Bytte till double

+++ Sources/HAP/Accessories/Energy.swift
@@ -1,10 +1,3 @@
-//
-//  Energy.swift
-//  HAP
-//
-//  Created by Thomas Gustafsson on 2017-10-23.
-//
-
 import Foundation
 
 extension Accessory {
@@ -21,8 +14,8 @@ extension Service {
     open class Energy: Service {
         public let on = GenericCharacteristic<Bool>(type: .on, value: false)
         public let inUse = GenericCharacteristic<Bool>(type: .outletInUse, value: true, permissions: [.read, .events])
-        public let watt = GenericCharacteristic<Int>(type: .watt, value: 0, permissions: [.read, .events])
-        public let kiloWattHour = GenericCharacteristic<Int>(type: .kiloWattHour, value: 0, permissions: [.read, .events])
+        public let watt = GenericCharacteristic<Double>(type: .watt, value: 0, permissions: [.read, .events])
+        public let kiloWattHour = GenericCharacteristic<Double>(type: .kiloWattHour, value: 0, permissions: [.read, .events])
 
         public init() {
             super.init(type: .outlet, characteristics: [on, inUse,watt,kiloWattHour])
